+++ spec/cmds/deploy_spec.js
@@ -199,13 +199,10 @@ h.describeRequireVm('Azk cli, deploy controller', function() {
 
     it("should no deploy system", function() {
       doc_opts.argv = ['deploy'];
-      var options = cli.router.cleanParams(cli.docopt(doc_opts));
       var run_options = { ui: ui, cwd: manifest.cwd, just_parse: true };
 
-      return cli.run(doc_opts, run_options).then((code) => {
-        h.expect(code).to.equal(1);
-        h.expect(options).to.have.property('deploy', true);
-      });
+      var result = cli.run(doc_opts, run_options);
+      return h.expect(result).to.be.rejectedWith(SystemNotFoundError, /deploy/);
     });
   });
 });