+++ spec/utils/tracker_spec.js
@@ -1,6 +1,5 @@
 import h from 'spec/spec_helper';
 import { default as tracker } from 'azk/utils/tracker';
-import { meta as azkMeta } from 'azk';
 
 describe("Azk Tracker", function() {
   this.timeout(2000);
@@ -12,7 +11,7 @@ describe("Azk Tracker", function() {
   it("should generateRandomId", function() {
     var label = 'agent_session_id';
     var new_hash = tracker.generateRandomId(label);
-    var expected = label + ':' + '12345678';
+    var expected = label + ':' + '123456789012345';
     h.expect(new_hash.length).to.equal(expected.length);
   });
 
@@ -22,11 +21,4 @@ describe("Azk Tracker", function() {
     h.expect(sessionId).to.equal(expectedSessionId);
   });
 
-  it("should override insight get/set for optout", function() {
-    return azkMeta.cleanAsync().then(() => {
-      tracker.insight.optOut = true;
-      var metaOptout = azkMeta.get(tracker.insight_opts.opt_out_key);
-      h.expect(metaOptout).to.equal(false);
-    });
-  });
 });