+++ src/cli/views/view.js
@@ -1,0 +1,4 @@
+import { UIProxy } from '../ui';
+
+export class View extends UIProxy {
+}