+++ src/manifest/index.js
@@ -361,7 +361,7 @@ export class Manifest {
     return path.relative(this.manifestPath, this.file);
   }
 
-  // TODO: make more fast
+  // TODO: make faster
   get namespace() {
     var def = Utils.calculateHash(this.file).slice(0, 10);
     return this.meta.getOrSet('namespace', def);